Populate AnnualBilling automatically each year
At the start of each financial year, let's automatically populate
AnnualBilling for all services that exist.

This will ensure that the org usage dashboard immediately reflects free
allowances for all of its services.

Newly-created services get an AnnualBilling entry added upon creation.
